refactor(launcher_config): move models into seperate crate
