+++ src-tauri/Cargo.toml
@@ -1,5 +1,5 @@
 [workspace]
-members = [".", "crates/sjmcl-macros", "crates/sjmcl-static", "crates/sjmcl-types"]
+members = [".", "crates/sjmcl-launcher-config", "crates/sjmcl-macros", "crates/sjmcl-static", "crates/sjmcl-types"]
 resolver = "2"
 
 [workspace.package]
@@ -52,6 +52,7 @@ sha1 = "0.10.1"
 sha2 = "0.10.9"
 shlex = "1.3.0"
 sjmcl_macros = { path = "crates/sjmcl-macros" }
+sjmcl_launcher_config = { package = "sjmcl-launcher-config", path = "crates/sjmcl-launcher-config" }
 sjmcl_static = { package = "sjmcl-static", path = "crates/sjmcl-static" }
 sjmcl_types = { package = "sjmcl-types", path = "crates/sjmcl-types" }
 smart-default = "0.7.1"
@@ -157,6 +158,7 @@ sha1.workspace = true
 sha2.workspace = true
 shlex.workspace = true
 sjmcl_macros.workspace = true
+sjmcl_launcher_config.workspace = true
 sjmcl_static.workspace = true
 sjmcl_types.workspace = true
 smart-default.workspace = true
